+++ packages/mdx/tsdown.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'tsdown';
 export default defineConfig({
   entry: [
     './src/{index,bin}.ts',
-    './src/{config,next,vite,bun}/index.ts',
+    './src/{config,next,vite,bun,rolldown}/index.ts',
     './src/webpack/{mdx,meta}.ts',
     './src/node/{index,_loader,loader}.ts',
     './src/runtime/*.{ts,tsx}',
@@ -16,7 +16,7 @@ export default defineConfig({
   target: 'es2023',
   platform: 'neutral',
   exports: {
-    bin: './src/bin.ts',
+    bin: false,
   },
   deps: {
     onlyBundle: ['@fumadocs/vite'],