+++ packages/preview/fumadocs.config.ts
@@ -1,0 +1,7 @@
+import type { AppConfig } from '@/config';
+
+export default {
+  content: {
+    unsafe_disableFileSystemChecks: true,
+  },
+} satisfies AppConfig;