+++ packages/preview/src/config/global.ts
@@ -18,7 +18,7 @@ export const layoutConfigSchema = z.object({
       /**
        * layout preset for Markdown files.
        */
-      md: z.literal(['docs', 'flux']).optional(),
+      md: z.literal(['docs', 'flux']).default('docs'),
     })
     .optional(),
 });
@@ -52,25 +52,33 @@ export const contentConfigSchema = z.object({
    * a list of project configurations.
    */
   projects: z.array(projectConfigSchema).optional(),
+
+  /**
+   * allow file-system access outside of project directories.
+   */
+  unsafe_disableFileSystemChecks: z.boolean().default(false),
 });
 
 export const aiConfigSchema = z.object({
   /**
-   * model ID
+   * Specify a model ID, by default it uses OpenRouter.
    */
   model: z.string().optional(),
+
+  ratelimit: z.boolean().default(true),
 });
 
 export const configSchema = z.object({
-  layout: layoutConfigSchema.optional(),
-  content: contentConfigSchema.optional(),
-  ai: aiConfigSchema.optional(),
+  layout: layoutConfigSchema.default(() => layoutConfigSchema.parse({})),
+  content: contentConfigSchema.default(() => contentConfigSchema.parse({})),
+  ai: aiConfigSchema.default(() => aiConfigSchema.parse({})),
 });
 
 export type LayoutConfig = z.infer<typeof layoutConfigSchema>;
 export type ContentConfig = z.infer<typeof contentConfigSchema>;
 export type ProjectConfig = z.infer<typeof projectConfigSchema>;
-export type AppConfig = z.infer<typeof configSchema>;
+export type AppConfig = z.input<typeof configSchema>;
+export type ParsedAppConfig = z.output<typeof configSchema>;
 
 export function defineConfig(config: AppConfig = {}): AppConfig {
   return config;