+++ packages/mdx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # next-docs-mdx
 
+## 15.0.5
+
+### Patch Changes
+
+- 1fb6a61: Support custom base directory for content sources
+
+## 15.0.4
+
+### Patch Changes
+
+- 819b6ec: Support Rolldown integration
+
 ## 15.0.3
 
 ### Patch Changes